+++ scripts/lint.sh
@@ -2,7 +2,7 @@
 # run various linters
 set -e
 echo "running ruff..."
+python -m ruff format . --check
 python -m ruff check .
-python -m ruff .
 echo "running mypy..."
 python3 -m mypy --show-error-codes yellowbox